move vendor back into src with a special name
